+++ .gitignore
@@ -7,3 +7,4 @@ venv/
 /.vscode
 .env
 *.env
+sample.txt
